fix(manager_versions.yaml): remove duplicate, leave only base distros
we keep updating this file, on every adding of a new distro release
it's not really needed, we have only two types of supported RPMs (rhel)
and DEBs (debian), and that's enough for picking the repo/list
we don't need to map each distro release, since we have it mapped
already.

at a point there were diffrent DEB repositories for scylla/manager
it was unified years back.